Fixed compiler warning.

+++ src/libxsmm_dnn_rnncell.c
@@ -876,7 +876,7 @@ LIBXSMM_API libxsmm_dnn_tensor_datalayout* libxsmm_dnn_rnncell_create_tensor_dat
 
 LIBXSMM_API size_t libxsmm_dnn_rnncell_get_scratch_size(const libxsmm_dnn_rnncell* handle, const libxsmm_dnn_compute_kind kind, libxsmm_dnn_err_t* status)
 {
-  const int typesize_in = (int)libxsmm_dnn_typesize(handle->desc.datatype_in);
+  const size_t typesize_in = libxsmm_dnn_typesize(handle->desc.datatype_in);
   const size_t dwdr_typesize = (handle->desc.datatype_out == LIBXSMM_DNN_DATATYPE_BF16) ? sizeof(float) : typesize_in;
   size_t size = 0;
   *status = LIBXSMM_DNN_SUCCESS;
@@ -1013,7 +1013,7 @@ LIBXSMM_API void* libxsmm_dnn_rnncell_get_scratch_ptr(const libxsmm_dnn_rnncell*
 LIBXSMM_API libxsmm_dnn_err_t libxsmm_dnn_rnncell_bind_scratch(libxsmm_dnn_rnncell* handle, const libxsmm_dnn_compute_kind kind, const void* scratch)
 {
   libxsmm_dnn_err_t status = LIBXSMM_DNN_SUCCESS;
-  const int typesize_in = (int)libxsmm_dnn_typesize(handle->desc.datatype_in);
+  const size_t typesize_in = libxsmm_dnn_typesize(handle->desc.datatype_in);
   const size_t dwdr_typesize = (handle->desc.datatype_out == LIBXSMM_DNN_DATATYPE_BF16) ? sizeof(float) : typesize_in;
   uintptr_t address = (uintptr_t)scratch;
   size_t offset = 0;
@@ -2291,3 +2291,4 @@ LIBXSMM_API libxsmm_dnn_err_t libxsmm_dnn_rnncell_execute_st(libxsmm_dnn_rnncell
 
   return status;
 }
+
